site/pyproject.toml
[project]
name = "site"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Rain", email = "rain@sunshowers.io" }
]
dependencies = [
    "mkdocs-material[imaging]>=9.5.25",
    "ansi2html>=1.9.1",
    "markdown-exec[ansi]>=1.8.3",
    "mdx-truly-sane-lists>=1.3",
    "mkdocs-git-revision-date-localized-plugin>=1.2.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
dev-dependencies = []
